backend/kabadiman/urls.py
"""
Kabadi Man — URL Configuration.
"""

from django.contrib import admin
from django.urls import include, path
from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    # Django Admin
    path("admin/", admin.site.urls),
    # API Schema
    path("api/schema/", SpectacularAPIView.as_view(), name="schema"),
    path("api/docs/", SpectacularSwaggerView.as_view(url_name="schema"), name="swagger-ui"),
    # App URLs
    path("api/v1/token/", TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path("api/v1/token/refresh/", TokenRefreshView.as_view(), name='token_refresh'),
    path("api/v1/", include("apps.cities.urls")),
    path("api/v1/auth/", include("apps.accounts.urls")),
    path("api/v1/", include("apps.pricing.urls")),
    path("api/v1/", include("apps.orders.urls")),
    path("api/v1/", include("apps.invoices.urls")),
    path("api/v1/", include("apps.wallet.urls")),
    path("api/v1/", include("apps.notifications.urls")),
]
