backend/requirements/base.txt
# ── Core ─────────────────────────────────────────────
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
django-cors-headers>=4.3,<5.0
drf-spectacular>=0.27,<1.0

# ── Database ─────────────────────────────────────────
psycopg2-binary>=2.9,<3.0

# ── Async Tasks ──────────────────────────────────────
celery>=5.4,<6.0
redis>=5.0,<6.0

# ── Media / Files ───────────────────────────────────
cloudinary>=1.40,<2.0
django-cloudinary-storage>=0.3,<1.0
Pillow>=10.0,<11.0

# ── Security ─────────────────────────────────────────
bcrypt>=4.1,<5.0

# ── HTTP ─────────────────────────────────────────────
requests>=2.31,<3.0
